+++ R/plotting.R
@@ -6,10 +6,10 @@
 #'   values to separate minor, moderate, and severe defoliation events. If
 #'   blank, the mean and 1st quartile are used.
 #'
-#' @importFrom rlang .data
+#' @importFrom rlang .data :=
 #' @importFrom ggplot2 ggplot aes geom_segment theme_bw theme element_blank
 #'
-#' @example
+#' @examples
 #' data("dmj_defol")
 #' plot_defol(dmj_defol)
 #'
@@ -66,70 +66,85 @@ plot_defol <- function(x, breaks) {
 #' Produce a stacked plot to present composited, site-level insect outbreak
 #' chronologies
 #'
-#' @param x an 'outbreak' object produced by [outbreak()]
+#' @param x an 'obr' object produced by [outbreak()]
 #' @param disp_index Identify the timeseries index to plot. Defaults to
-#'   `mean_ngsi`, the average normalized growth suppression index for the
-#'   site. The only other option is `mean_gsi`, the average growth
+#'   `NGSI`, the average normalized growth suppression index for the
+#'   site. The only other option is `GSI`, the average growth
 #'   suppression index.
 #'
 #' @importFrom rlang .data
+#' @importFrom  magrittr %>%
 #' @importFrom ggplot2 ggplot aes geom_segment geom_vline theme_bw theme
 #'   element_blank geom_line geom_hline scale_y_continuous scale_x_continuous
 #'   geom_ribbon unit
 #'
 #' @export
-plot_outbreak <- function(x, disp_index = "mean_ngsi") {
+plot_outbreak <- function(x, disp_index = c("GSI", "NGSI")){
+
   if (!is.obr(x)) stop("'x' must be an 'obr' object")
-  if (! (disp_index == "mean_gsi" | disp_index == "mean_ngsi")) {
-    disp_index <- "mean_ngsi"
+  if (missing(disp_index)) disp_index <- "NGSI"
+  if (! disp_index %in% c("NGSI", "GSI")){
+    stop("Please assign either 'NGSI' or 'GSI' to the `disp_index` argument")
   }
-  if (disp_index == "mean_ngsi") y_intercept <- 0
-  else y_intercept <- 1
-  outbrk_events <- x[! x$outbreak_status %in% "not_obr", ]
+
+  if (disp_index == "NGSI"){
+    y_intercept <- 0
+    var <- "mean_ngsi"
+  }
+  if (disp_index == "GSI"){
+    y_intercept <- 1
+    var <- "mean_gsi"
+  }
+
+  outbrk_events <- x %>%
+    dplyr::mutate(!!var := replace(.data[[var]],
+                                   .data$outbreak_status == "not_obr",
+                                   y_intercept))
 
   # setup plot
   p <- ggplot(data = x, aes(x = .data$year))
   # extract minor axes
-  foo <- p + geom_line(aes(y = .data[[disp_index]]))
+  foo <- p + geom_line(aes(y = .data[[var]]))
   minor_labs <-
-    ggplot2::ggplot_build(foo)$layout$panel_params[[1]]$x.minor_source
+    ggplot2::ggplot_build(foo)$layout$panel_params[[1]]$x.sec$minor_breaks
   # top plot
-  index <- p +
-    geom_vline(xintercept = minor_labs, colour = "grey50") +
-    geom_hline(yintercept = y_intercept, colour = "grey80") +
-    geom_line(aes(y = .data[[disp_index]])) +
-    geom_segment(data = outbrk_events,
-                          aes(x = .data$year,
-                              xend = .data$year,
-                              y = y_intercept,
-                              yend = .data[[disp_index]]),
-                          size = 2) +
-    scale_y_continuous(name = "GSI") +
+  line <- p +
+    geom_vline(xintercept = minor_labs, colour = "grey90") +
+    geom_line(aes(y = .data$samp_depth)) +
+    scale_y_continuous(name = "# trees") +
     ggpubr::theme_pubr() +
     theme(plot.margin = unit(c(0.1, 0, 0, 0), "cm"),
-                   axis.title.x = element_blank(),
-                   axis.text.x = element_blank(),
-                   axis.ticks.x = element_blank())
-  # mid plot
-  prop <- p +
-    geom_vline(xintercept = minor_labs, colour = "grey50") +
-    geom_ribbon(aes(ymax = .data$perc_defol, ymin = 0)) +
-    scale_y_continuous(name = "% defoliated") +
+          axis.title.x = element_blank(),
+          axis.text.x = element_blank(),
+          axis.ticks.x = element_blank())
+  # middle plot
+  index <- p +
+    geom_vline(xintercept = minor_labs, colour = "grey90") +
+    geom_hline(yintercept = y_intercept, colour = "grey80") +
+    # geom_area(data=outbrk_events, aes(x = .data$year,
+    #                                   y = .data[[var]]),
+    #           stat = "identity") +
+    geom_ribbon(data=outbrk_events, aes(x = .data$year,
+                                        ymin = .data[[var]],
+                                        ymax = y_intercept)) +
+    geom_line(aes(y = .data[[var]])) +
+    scale_y_continuous(name = disp_index) +
     ggpubr::theme_pubr() +
     theme(plot.margin = unit(c(0.1, 0, 0, 0), "cm"),
                    axis.title.x = element_blank(),
                    axis.text.x = element_blank(),
                    axis.ticks.x = element_blank())
   # bottom plot
-  line <- p +
-    geom_vline(xintercept = minor_labs, colour = "grey50") +
-    geom_line(aes(y = .data$samp_depth)) +
-    scale_y_continuous(name = "Sample depth") +
+  prop <- p +
+    geom_vline(xintercept = minor_labs, colour = "grey90") +
+    geom_ribbon(aes(ymax = .data$perc_defol, ymin = 0)) +
+    scale_y_continuous(name = "% defoliated") +
     scale_x_continuous(name = "Year") +
     ggpubr::theme_pubr() +
     theme(plot.margin = unit(c(0.1, 0, 0, 0), "cm"))
+
   # combine
-  ggpubr::ggarrange(index, prop, line, nrow = 3, align = "v")
+  ggpubr::ggarrange(line, index, prop, nrow = 3, align = "v", heights = c(1, 2, 2))
 }
 
 #' Plot a `defol` object