+++ NEWS.md
@@ -5,6 +5,7 @@
 * Adding new logo
 * Fixed a bug in `id_defoliation()` regarding series_end_events
 * Added updated East Fork data, new objects are `efk_*`
+* Fix up `plot_outbreak()` with flexible Y-axis labels and shaed area
 
 # dfoliatR v0.1.0